Fixed stringcyclers
CoreRPG updated stringcyclers a few weeks ago to call update through an onUpdate DB handler. This meant that our update() function was getting a databasenode passed to it and not the boolean it expected. So now we have to check if the argument is a boolean or not and process accordingly.

A better solution might have been to make the K&W function a different name, but every other window control uses an update(bool, bool) function to show/hide itself. This is a case of CoreRPG stringcycler breaking that rule.